refactors suggestion controller out of story controller
