+++ app/helpers/suggestions_helper.rb
@@ -1,0 +1,2 @@
+module SuggestionsHelper
+end
