+++ app/views/suggestions/create.html.erb
@@ -1,0 +1,2 @@
+<h1>Suggestions#create</h1>
+<p>Find me in app/views/suggestions/create.html.erb</p>
